+++ capability_outputs.tf
@@ -15,5 +15,9 @@ locals {
     "deployment_annotations",
     "service_annotations",
     "deployment_overrides",
+    "resource_limits",
+    "node_selectors",
+    "tolerations",
+    "topology_spread_constraints",
   ]
 }